decoupled api classes
