+++ tests/api/test_splits_api.py
@@ -4,8 +4,10 @@ import pytest
 import unittest.mock as mock
 import time
 
-from splitio.api import splits, client, APIException
-from splitio.api.commons import FetchOptions
+from splitio.spec import SPEC_VERSION
+from splitio.api import splits
+from harness_commons.api import client, APIException
+from harness_commons.api.commons import FetchOptions
 from splitio.client.util import SdkMetadata
 
 class SplitAPITests(object):
@@ -17,7 +19,7 @@ class SplitAPITests(object):
         httpclient.get.return_value = client.HttpResponse(200, '{"prop1": "value1"}', {})
         split_api = splits.SplitsAPI(httpclient, 'some_api_key', SdkMetadata('1.0', 'some', '1.2.3.4'), mocker.Mock())
 
-        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, 'set1,set2'))
+        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, 'set1,set2', SPEC_VERSION))
         assert response['prop1'] == 'value1'
         assert httpclient.get.mock_calls == [mocker.call('sdk', 'splitChanges', 'some_api_key',
                                                          extra_headers={
@@ -28,7 +30,7 @@ class SplitAPITests(object):
                                                          query={'s': '1.3', 'since': 123, 'rbSince': -1, 'sets': 'set1,set2'})]
 
         httpclient.reset_mock()
-        response = split_api.fetch_splits(123, 1, FetchOptions(True, 123, None,'set3'))
+        response = split_api.fetch_splits(123, 1, FetchOptions(True, 123, None,'set3', SPEC_VERSION))
         assert response['prop1'] == 'value1'
         assert httpclient.get.mock_calls == [mocker.call('sdk', 'splitChanges', 'some_api_key',
                                                          extra_headers={
@@ -40,7 +42,7 @@ class SplitAPITests(object):
                                                          query={'s': '1.3', 'since': 123, 'rbSince': 1, 'till': 123, 'sets': 'set3'})]
 
         httpclient.reset_mock()
-        response = split_api.fetch_splits(123, 122, FetchOptions(True, 123, None, 'set3'))
+        response = split_api.fetch_splits(123, 122, FetchOptions(True, 123, None, 'set3', SPEC_VERSION))
         assert response['prop1'] == 'value1'
         assert httpclient.get.mock_calls == [mocker.call('sdk', 'splitChanges', 'some_api_key',
                                                          extra_headers={
@@ -78,7 +80,7 @@ class SplitAPITests(object):
 
         httpclient.is_sdk_endpoint_overridden.return_value = False
         try: 
-            response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+            response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         except Exception as e:
             print(e)
             
@@ -88,7 +90,7 @@ class SplitAPITests(object):
         httpclient.is_sdk_endpoint_overridden.return_value = True
         self.query = []
         self.counter = 0
-        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))        
+        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))        
         assert response == {"ff": {"d": [], "s": 123, "t": 456}, "rbs": {"d": [], "s": -1, "t": -1}}
         assert self.query == [{'s': '1.3', 'since': 123, 'rbSince': -1}, {'s': '1.1', 'since': 123}]
         assert not split_api.clear_storage
@@ -111,14 +113,14 @@ class SplitAPITests(object):
         httpclient.is_sdk_endpoint_overridden.return_value = True
         httpclient.get = get
         split_api = splits.SplitsAPI(httpclient, 'some_api_key', SdkMetadata('1.0', 'some', '1.2.3.4'), mocker.Mock())
-        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         assert response == {"ff": {"d": [], "s": 123, "t": 456}, "rbs": {"d": [], "s": -1, "t": -1}}
         assert self.query == [{'s': '1.3', 'since': 123, 'rbSince': -1}, {'s': '1.1', 'since': 123}]
         assert not split_api.clear_storage
 
         time.sleep(1)
         splits._PROXY_CHECK_INTERVAL_MILLISECONDS_SS = 10
-        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         assert self.query[2] == {'s': '1.3', 'since': 123, 'rbSince': -1}
         assert response == {"ff": {"d": [], "s": 123, "t": 456}, "rbs": {"d": [], "s": 123, "t": -1}}
         assert split_api.clear_storage
@@ -143,7 +145,7 @@ class SplitAPITests(object):
         httpclient.is_sdk_endpoint_overridden.return_value = True
         httpclient.get = get
         split_api = splits.SplitsAPI(httpclient, 'some_api_key', SdkMetadata('1.0', 'some', '1.2.3.4'), mocker.Mock())
-        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+        response = split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         assert response == {"ff": {"d": [], "s": 123, "t": 456}, "rbs": {"d": [], "s": -1, "t": -1}}
         assert self.query == [{'s': '1.3', 'since': 123, 'rbSince': -1}, {'s': '1.1', 'since': 123}]
         assert not split_api.clear_storage
@@ -151,7 +153,7 @@ class SplitAPITests(object):
         time.sleep(1)
         splits._PROXY_CHECK_INTERVAL_MILLISECONDS_SS = 10
         
-        response = split_api.fetch_splits(456, -1, FetchOptions(False, None, None, None))
+        response = split_api.fetch_splits(456, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         time.sleep(1)        
         splits._PROXY_CHECK_INTERVAL_MILLISECONDS_SS = 1000000
         assert self.query[2] == {'s': '1.3', 'since': 456, 'rbSince': -1}
@@ -180,7 +182,7 @@ class SplitAPIAsyncTests(object):
             return client.HttpResponse(200, '{"prop1": "value1"}', {})
         httpclient.get = get
 
-        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, 'set1,set2'))
+        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, 'set1,set2', SPEC_VERSION))
         assert response['prop1'] == 'value1'
         assert self.verb == 'sdk'
         assert self.url == 'splitChanges'
@@ -193,7 +195,7 @@ class SplitAPIAsyncTests(object):
         assert self.query == {'s': '1.3', 'since': 123, 'rbSince': -1, 'sets': 'set1,set2'}
 
         httpclient.reset_mock()
-        response = await split_api.fetch_splits(123, 1, FetchOptions(True, 123, None, 'set3'))
+        response = await split_api.fetch_splits(123, 1, FetchOptions(True, 123, None, 'set3', SPEC_VERSION))
         assert response['prop1'] == 'value1'
         assert self.verb == 'sdk'
         assert self.url == 'splitChanges'
@@ -207,7 +209,7 @@ class SplitAPIAsyncTests(object):
         assert self.query == {'s': '1.3', 'since': 123, 'rbSince': 1, 'till': 123, 'sets': 'set3'}
 
         httpclient.reset_mock()
-        response = await split_api.fetch_splits(123, 122, FetchOptions(True, 123, None))
+        response = await split_api.fetch_splits(123, 122, FetchOptions(True, 123, None, None, SPEC_VERSION))
         assert response['prop1'] == 'value1'
         assert self.verb == 'sdk'
         assert self.url == 'splitChanges'
@@ -249,7 +251,7 @@ class SplitAPIAsyncTests(object):
 
         httpclient.is_sdk_endpoint_overridden.return_value = False
         try: 
-            response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+            response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         except Exception as e:
             print(e)
             
@@ -259,7 +261,7 @@ class SplitAPIAsyncTests(object):
         httpclient.is_sdk_endpoint_overridden.return_value = True
         self.query = []
         self.counter = 0
-        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         assert response == {"ff": {"d": [], "s": 123, "t": 456}, "rbs": {"d": [], "s": -1, "t": -1}}
         assert self.query == [{'s': '1.3', 'since': 123, 'rbSince': -1}, {'s': '1.1', 'since': 123}]
         assert not split_api.clear_storage
@@ -283,14 +285,14 @@ class SplitAPIAsyncTests(object):
         httpclient.is_sdk_endpoint_overridden.return_value = True
         httpclient.get = get
         split_api = splits.SplitsAPIAsync(httpclient, 'some_api_key', SdkMetadata('1.0', 'some', '1.2.3.4'), mocker.Mock())
-        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         assert response == {"ff": {"d": [], "s": 123, "t": 456}, "rbs": {"d": [], "s": -1, "t": -1}}
         assert self.query == [{'s': '1.3', 'since': 123, 'rbSince': -1}, {'s': '1.1', 'since': 123}]
         assert not split_api.clear_storage
 
         time.sleep(1) 
         splits._PROXY_CHECK_INTERVAL_MILLISECONDS_SS = 10
-        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         assert self.query[2] == {'s': '1.3', 'since': 123, 'rbSince': -1}
         assert response == {"ff": {"d": [], "s": 123, "t": 456}, "rbs": {"d": [], "s": 123, "t": -1}}
         assert split_api.clear_storage
@@ -316,7 +318,7 @@ class SplitAPIAsyncTests(object):
         httpclient.is_sdk_endpoint_overridden.return_value = True
         httpclient.get = get
         split_api = splits.SplitsAPIAsync(httpclient, 'some_api_key', SdkMetadata('1.0', 'some', '1.2.3.4'), mocker.Mock())
-        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None))
+        response = await split_api.fetch_splits(123, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         assert response == {"ff": {"d": [], "s": 123, "t": 456}, "rbs": {"d": [], "s": -1, "t": -1}}
         assert self.query == [{'s': '1.3', 'since': 123, 'rbSince': -1}, {'s': '1.1', 'since': 123}]
         assert not split_api.clear_storage
@@ -324,7 +326,7 @@ class SplitAPIAsyncTests(object):
         time.sleep(1)
         splits._PROXY_CHECK_INTERVAL_MILLISECONDS_SS = 10
         
-        response = await split_api.fetch_splits(456, -1, FetchOptions(False, None, None, None))
+        response = await split_api.fetch_splits(456, -1, FetchOptions(False, None, None, None, SPEC_VERSION))
         time.sleep(1)        
         splits._PROXY_CHECK_INTERVAL_MILLISECONDS_SS = 1000000
         assert self.query[2] == {'s': '1.3', 'since': 456, 'rbSince': -1}
