+++ tests/sync/test_events_synchronizer.py
@@ -4,8 +4,8 @@ import threading
 import time
 import pytest
 
-from splitio.api.client import HttpResponse
-from splitio.api import APIException
+from harness_commons.api.client import HttpResponse
+from harness_commons.api import APIException
 from splitio.storage import EventStorage
 from harness_commons.models.events import Event
 from splitio.sync.event import EventSynchronizer, EventSynchronizerAsync
