+++ tests/sync/test_impressions_synchronizer.py
@@ -4,8 +4,8 @@ import threading
 import time
 import pytest
 
-from splitio.api.client import HttpResponse
-from splitio.api import APIException
+from harness_commons.api.client import HttpResponse
+from harness_commons.api import APIException
 from splitio.storage import ImpressionStorage
 from harness_commons.models.impressions import Impression
 from splitio.sync.impression import ImpressionSynchronizer, ImpressionSynchronizerAsync
