+++ tests/tasks/test_impressions_sync.py
@@ -4,11 +4,11 @@ import threading
 import time
 import pytest
 
-from splitio.api.client import HttpResponse
+from harness_commons.api.client import HttpResponse
 from splitio.tasks import impressions_sync
 from splitio.storage import ImpressionStorage
 from harness_commons.models.impressions import Impression
-from splitio.api.impressions import ImpressionsAPI
+from harness_commons.api.impressions import ImpressionsAPI
 from splitio.sync.impression import ImpressionSynchronizer, ImpressionsCountSynchronizer, ImpressionSynchronizerAsync, ImpressionsCountSynchronizerAsync
 from splitio.engine.impressions.manager import Counter
 from splitio.optional.loaders import asyncio
